fix(server): storage usage should be float rather than int (#4275)

+++ apps/server/src/modules/workspaces/resolver.ts
@@ -101,7 +101,7 @@ export class InvitationWorkspaceType {
 
 @ObjectType()
 export class WorkspaceBlobSizes {
-  @Field(() => Int)
+  @Field(() => Float)
   size!: number;
 }
